Add support for "tech cookie"
Use "tech cookie" to determine whether the client making
requests is abled to store cookies. On every request presence and age of
the "tech cookie" will be checked. The CMP status will be disabled in
case the "tech cookie" is missing or doesn't have the minimum age (by
default two days). The "tech cookie" stores no data used for
identification.