Add support for additional channels banner
This change implements an additional banner which can be used to request
the user to confirm their consent once the tracking channel list
changes.

To allow the integrator of consent management API to choose between
banners to show an additional API method has been added:
`showAdditionalChannelsBanner`, this method behaves exactly the same way
as `showBanner` with the exception that the callback function is called
with two parameters when decline is selected. The second parameter
provides the action name (reject, settings).

TVI-1943